+++ app-backend/batch/src/main/scala/groundwork/NotifyIntercom.scala
@@ -6,40 +6,30 @@ import com.rasterfoundry.notification.intercom.IntercomConversation
 import com.rasterfoundry.notification.intercom.Model._
 import com.rasterfoundry.notification.intercom._
 
-import cats.effect.{Async, IO}
-import sttp.client.asynchttpclient.cats.AsyncHttpClientCatsBackend
+import cats.effect.IO
+import sttp.client3.asynchttpclient.cats.AsyncHttpClientCatsBackend
 
 object NotifyIntercomProgram extends Job {
   val name = "notify-intercom"
-
-  val getBackend = for {
-    backendRef <- Async.memoize {
-      AsyncHttpClientCatsBackend[IO]()
-    }
-    backend <- backendRef
-  } yield backend
 
   def runJob(args: List[String]): IO[Unit] = {
     val xa =
       RFTransactor.nonHikariTransactor(RFTransactor.TransactorConfig())
 
-    val dbIO = new DbIO(xa);
+    val dbIO = new DbIO(xa)
 
     args match {
       case externalId +: msg +: Nil =>
-        for {
-          backend <- getBackend
-          _ <- {
-            IntercomConversation.notifyIO(
-              externalId,
-              Message(msg),
-              dbIO.groundworkConfig,
-              new LiveIntercomNotifier[IO](backend),
-              dbIO.getConversation,
-              dbIO.insertConversation
-            )
-          }
-        } yield ()
+        AsyncHttpClientCatsBackend.resource[IO]() use { backend =>
+          IntercomConversation.notifyIO(
+            externalId,
+            Message(msg),
+            dbIO.groundworkConfig,
+            new LiveIntercomNotifier[IO](backend),
+            dbIO.getConversation,
+            dbIO.insertConversation
+          )
+        }
       case _ =>
         IO.raiseError(
           new Exception(